Make Manager Singleton and Add configuration

+++ Sources/CTCoreData/CoreDataManager.swift
@@ -5,15 +5,22 @@
 //  Created by Cem Telliagaoglu on 17.04.2024.
 //
 
-import Foundation
+import UIKit
 import CoreData
 
 open class CoreDataManager {
     // MARK: - Properties
     
-    private let storageName: String
+    public static let shared = CoreDataManager()
+    
+    private var isConfigured = false
+    private var storageName: String?
 
     private lazy var persistentContainer: NSPersistentContainer = {
+        guard let storageName else {
+            displayError(.storage)
+            return NSPersistentContainer(name: "")
+        }
         let container = NSPersistentContainer(name: storageName)
         container.loadPersistentStores(completionHandler: { _, error in
             if let error = error as NSError? {
@@ -27,11 +34,13 @@ open class CoreDataManager {
 
     // MARK: - Lifecycle
     
-    public init(storageName: String) {
-        self.storageName = storageName
-    }
+    private init() {}
     
     //MARK: - Methods
+    
+    open func configureDataModel(storageName: String) {
+        self.storageName = storageName
+    }
 
     open func create<E>(type _: E.Type, completion: @escaping ((Result<E, CoreDataError>) -> Void)) {
         do {
@@ -77,12 +86,28 @@ open class CoreDataManager {
     open func delete(objects: [NSManagedObject], completion: @escaping ((Result<Void, CoreDataError>) -> Void)) {
         do {
             for object in objects {
+                guard object.managedObjectContext == context else {
+                    completion(.failure(.storage))
+                    return
+                }
                 context.delete(object)
             }
             try context.save()
             completion(.success(()))
         } catch {
             completion(.failure(.delete))
+        }
+    }
+    
+    private func displayError(_ error: CoreDataError) {
+        let alert = UIAlertController(title: "Error", message: error.customMessage, preferredStyle: .alert)
+        alert.addAction(.init(title: "Close", style: .cancel))
+        if #available(iOS 13.0, *) {
+            if let windowScene = UIApplication.shared.connectedScenes.first as? UIWindowScene {
+                windowScene.windows.first?.rootViewController?.present(alert, animated: true)
+            }
+        } else {
+            UIApplication.shared.keyWindow?.rootViewController?.present(alert, animated: true)
         }
     }
 }
